refactor(tools_): relocate GAT helpers from enhance/ to tools_/
Move the private ``_anscombe.py`` math helpers from
``phenotypic.enhance`` to ``phenotypic.tools_`` (peer to ``funcs_.py``).
The helpers are pure utility functions with no enhancer-specific
state, so the prior location was conceptually misleading.

Side benefit: the relocation removes the circular-import workaround
in ``_GATSupportMixin``. The mixin's import of ``gat_forward``/
``gat_inverse``/``resolve_scale_factor`` can now live at module
top-level instead of inside ``_gat_apply``, since
``phenotypic.tools_`` does not pull in the enhance subpackage.

Also moves ``tests/unit/enhance/test_anscombe_helpers.py`` ->
``tests/unit/tools_/test_anscombe_helpers.py`` to mirror the new
source layout. Updates ``StableDenoise`` and the mixin tests to the
new import path.
